Clean up WIP commit
