+++ utils/contextMenu/calculateMenuHeight.ts
@@ -1,14 +1,5 @@
-import { contextMenuStyleGuide } from "./contstants";
-import { menuItemHeight } from "./menuItemHeight";
+import { MENU_ITEM_HEIGHT } from "./constants";
 
-export const calculateMenuHeight = (
-  itemLength: number,
-  separatorCount: number
-) => {
-  "worklet";
-  return (
-    menuItemHeight() * itemLength +
-    (itemLength - 1) +
-    separatorCount * contextMenuStyleGuide.spacing
-  );
+export const calculateMenuHeight = (itemLength: number) => {
+  return MENU_ITEM_HEIGHT * itemLength;
 };
